Implement MemberEntity.
Fix issues with repository creation

+++ backend/src/main/java/com/cmpt373sedna/gitlabanalyzer/repository/CommitEntityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CommitEntityRepository extends CrudRepository<CommitEntity, Integer> {
+public interface CommitEntityRepository extends CrudRepository<CommitEntity, Long> {
 
     @Query("SELECT c FROM CommitEntity c WHERE c.author = ?1")
     List<CommitEntity> getUserCommits(String author);
